+++ source-code/clj/server_extensions/clients/client_editor/views.clj
@@ -8,7 +8,7 @@
 ;; -- Lifecycle events --------------------------------------------------------
 ;; ----------------------------------------------------------------------------
 
-(a/reg-lifecycles! 
+(a/reg-lifecycles!
   ::lifecycles
   {:on-server-boot [:item-editor/initialize-editor! :clients :client
-                                                    {}]});:suggestion-keys [:city]}]})
+                                                    {:suggestion-keys [:city]}]})