+++ source-code/cljs/app_plugins/item_lister/queries.cljs
@@ -6,7 +6,7 @@
 ; Created: 2021.11.21
 ; Description:
 ; Version: v0.7.6
-; Compatibility: x4.6.0
+; Compatibility: x4.6.1
 
 
 
@@ -34,16 +34,16 @@
   ;  {:downloaded-item-count (integer)
   ;   :download-limit (integer)
   ;   :filter-pattern (maps in vector)
-  ;   :order-by (keyword)
+  ;   :order-by (namespaced keyword)
   ;   :reload-items? (boolean)
   ;   :search-keys (keywords in vector)
   ;   :search-term (string)}
   [db [_ extension-id item-namespace]]
-  {:downloaded-item-count (r subs/get-downloaded-item-count db extension-id item-namespace)
-   :download-limit        (r subs/get-meta-item             db extension-id item-namespace :download-limit)
+  {:download-limit        (r subs/get-meta-item             db extension-id item-namespace :download-limit)
    :order-by              (r subs/get-meta-item             db extension-id item-namespace :order-by)
    :reload-items?         (r subs/get-meta-item             db extension-id item-namespace :reload-mode?)
    :search-keys           (r subs/get-meta-item             db extension-id item-namespace :search-keys)
+   :downloaded-item-count (r subs/get-downloaded-item-count db extension-id item-namespace)
    :filter-pattern        (r subs/get-filter-pattern        db extension-id item-namespace)
    :search-term           (r subs/get-search-term           db extension-id item-namespace)