+++ source-code/cljs/app_plugins/item_lister/subs.cljs
@@ -5,8 +5,8 @@
 ; Author: bithandshake
 ; Created: 2021.11.21
 ; Description:
-; Version: v0.8.0
-; Compatibility: x4.6.0
+; Version: v0.8.4
+; Compatibility: x4.6.1